requirements.txt
torch~=1.12.1
Pillow>=7.1.2
easyocr==1.6.2
matplotlib==3.6.0
numpy~=1.23.3
pyyaml~=6.0
torchvision~=0.13.1
tqdm~=4.64.1
opencv-python~=4.6.0.66
pandas~=1.5.0
seaborn~=0.12.0
requests~=2.28.1
thop~=0.1.1-2209072238
pybind11==2.9.2
fastwer==0.1.3
